Faas 3: tõsta lehekülgede haldus routerisse #36
